Create a 'connection' packages at the API level and implementation level
