+++ api/concept/thing/Relation.java
@@ -21,7 +21,7 @@
 
 package com.vaticle.typedb.client.api.concept.thing;
 
-import com.vaticle.typedb.client.api.TypeDBTransaction;
+import com.vaticle.typedb.client.api.connection.TypeDBTransaction;
 import com.vaticle.typedb.client.api.concept.type.RelationType;
 import com.vaticle.typedb.client.api.concept.type.RoleType;
