+++ api/logic/Rule.java
@@ -21,7 +21,7 @@
 
 package com.vaticle.typedb.client.api.logic;
 
-import com.vaticle.typedb.client.api.TypeDBTransaction;
+import com.vaticle.typedb.client.api.connection.TypeDBTransaction;
 import com.vaticle.typeql.lang.pattern.Pattern;
 
 import javax.annotation.CheckReturnValue;
