+++ concept/thing/EntityImpl.java
@@ -21,7 +21,7 @@
 
 package com.vaticle.typedb.client.concept.thing;
 
-import com.vaticle.typedb.client.api.TypeDBTransaction;
+import com.vaticle.typedb.client.api.connection.TypeDBTransaction;
 import com.vaticle.typedb.client.api.concept.thing.Entity;
 import com.vaticle.typedb.client.concept.type.EntityTypeImpl;
 import com.vaticle.typedb.common.collection.Bytes;
