+++ concept/type/RelationTypeImpl.java
@@ -21,7 +21,7 @@
 
 package com.vaticle.typedb.client.concept.type;
 
-import com.vaticle.typedb.client.api.TypeDBTransaction;
+import com.vaticle.typedb.client.api.connection.TypeDBTransaction;
 import com.vaticle.typedb.client.api.concept.type.RelationType;
 import com.vaticle.typedb.client.common.Label;
 import com.vaticle.typedb.client.concept.thing.RelationImpl;
