+++ connection/BUILD
@@ -24,7 +24,7 @@ package(default_visibility = ["//visibility:public"])
 load("@vaticle_dependencies//tool/checkstyle:rules.bzl", "checkstyle_test")
 
 java_library(
-    name = "rpc",
+    name = "connection",
     srcs = glob(["*.java", "*/*.java"], exclude=["test/**"]),
     deps = [
         # Internal dependencies
