+++ test/behaviour/connection/ConnectionStepsCore.java
@@ -22,7 +22,7 @@
 package com.vaticle.typedb.client.test.behaviour.connection;
 
 import com.vaticle.typedb.client.TypeDB;
-import com.vaticle.typedb.client.api.TypeDBClient;
+import com.vaticle.typedb.client.api.connection.TypeDBClient;
 import com.vaticle.typedb.common.test.server.TypeDBCoreRunner;
 import com.vaticle.typedb.common.test.server.TypeDBSingleton;
 import io.cucumber.java.After;
